Module cardputer::keyboard

Keyboard scanner that converts the 74HC138 decoding results into the Vector of pressed key codes.

4x14 keymap

 A2 A2 A0 |    Y0  |  Y1   |  Y2   |  Y3   |  Y4   |  Y5   |  Y6
 H  -  -  |  L     | L     | L     | L     | L     | L     | L
 L  -  -  |      L |     L |     L |     L |     L |     L |     L
----------+--------+-------+-------+-------+-------+-------+-------
 -  H  H  |  `   1   2   3   4   5   6   7   8   9   0   -   =  DEL
 -  H  L  | TAB  q   w   e   r   t   y   u   i   o   p   [   ]   \
 -  L  H  | FN  SHT  a   s   d   f   g   h   j   k   l   ;   '  ENT
 -  L  L  | CTL OPT ALT  z   x   c   v   b   n   m   ,   .   /  SPC

Structs

ConversionRule

Conversion rule

Keyboard

Keyboard scanner for Cardputer

KeyboardState

Structure that scans the keyboard and keeps track of state changes

Enums

KeyImprint

KeyType

Define the type of key as modifier key and normal key

Modified

Traits

KeyboardScanner

Keyboard scanner trait